test: dedupe sanitize-transcript tests into canonical __tests__ location
PR 3's fix round created a second, more complete test file in the
providers/ root; consolidate it into providers/__tests__/ (correct
import paths) and remove the duplicate.
